requirements.txt
fastapi
uvicorn
python-dotenv
PyMuPDF
sentence-transformers
faiss-cpu
numpy
tqdm
pydantic
openai
transformers
torch
streamlit